major bug fix
Fixed a bug that caused crash when a user that didn't host tries to
unhost

+++ src/listeners/messageCreate.js
@@ -68,6 +68,10 @@ function hostCommand(command, hostingChannel, hostMap, joinEmoji)
 function unhostCommand(hostID, hostMap)
 {
 	const host = hostMap.get(hostID);
+	
+	if(host === undefined)
+		return;
+
 	hostMap.delete(hostID);
 	host.message.delete().catch(console.error);
 }
